requirements.txt
bcrypt==4.2.1
blinker==1.8.2
click==8.1.7
dnspython==2.7.0
Flask==3.0.3
Flask-Cors==5.0.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.0
mpmath==1.3.0
packaging==24.2
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sympy==1.13.3
Werkzeug==3.0.4
